Squashed commit of the following:
commit 77bc09210be674c023061eea4ac52d4657ff1df4
Merge: a5de8b4 34060ca
Author: zhxina <285733286@qq.com>
Date:   Mon Mar 18 15:59:16 2019 +0800

    Merge branch 'master' into zjh

commit a5de8b4faffb890d9edb94eabbb391ce5daeae20
Author: zjh <516413718@qq.com>
Date:   Mon Mar 18 15:46:00 2019 +0800

    新增用户管理

+++ src/com/freeway/dao/SlMockUserDao.java
@@ -4,7 +4,9 @@ import org.springframework.stereotype.Repository;
 
 import com.bstek.dorado.hibernate.HibernateDao;
 import com.freeway.entity.SlMockUser;
+
+
 @Repository
-public class SlMockUserDao extends HibernateDao<SlMockUser, Integer> {
+public class SlMockUserDao  extends HibernateDao<SlMockUser, Long>  {
 
 }
